Factor out remaining functions except for *Updater structs
